requirements.txt
emnist
numpy
matplotlib 
scikit-learn
Keras
Pillow